feat: 新增全局系统提示词功能 + 修复浮点精度丢失
新功能：全局系统提示词 (Global System Prompt)
- 在设置页面「高级设置」tab 中新增全局系统提示词配置
- 支持启用/禁用开关和多行文本编辑
- 提示词自动注入到所有 API 请求的 systemInstruction 中
- 注入位置：Antigravity 身份之后、客户端提示词之前
- 三种协议映射器 (OpenAI/Claude/Gemini) 均已适配
- 支持热更新：保存设置后立即生效，无需重启服务
- 中英文国际化翻译

修复：temperature/top_p 浮点精度丢失
- 将 OpenAI 和 Claude 模型的 temperature、top_p 从 f32 改为 f64
- 修复 top_p=0.95 序列化为 0.949999988079071 的精度问题

+++ src-tauri/src/proxy/mappers/openai/models.rs
@@ -16,9 +16,9 @@ pub struct OpenAIRequest {
     pub n: Option<u32>, // [NEW] 支持多候选结果数量
     #[serde(rename = "max_tokens")]
     pub max_tokens: Option<u32>,
-    pub temperature: Option<f32>,
+    pub temperature: Option<f64>,
     #[serde(rename = "top_p")]
-    pub top_p: Option<f32>,
+    pub top_p: Option<f64>,
     pub stop: Option<Value>,
     pub response_format: Option<ResponseFormat>,
     #[serde(default)]
